Separates trait and implementation into different files
