+++ consensus/scp/src/node/mod.rs
@@ -1,0 +1,9 @@
+// Copyright (c) 2018-2020 MobileCoin Inc.
+
+//! A node of the consensus network.
+
+mod node_impl;
+mod node_trait;
+
+pub use node_impl::*;
+pub use node_trait::*;
